Update Target_Detection
Ok it works now. Well, it doesn't have errors anymore. So basically to use this put the function and needed variables into the program you want and the function will spit out the y value of the target. ALSO VERY IMPORTANT: This program is kinda bad, so for best functionality use in a room without anything red in it or else the robot will be all outta wack